Add prometheus metrics for feeds
