+++ requirements.txt
@@ -15,3 +15,4 @@ urllib3==1.22
 Werkzeug==0.12.2
 yamlordereddictloader==0.4.0
 pyyaml==3.12
+prometheus-flask-exporter==0.2.0
